tentativa de consertar a Home
