+++ labe-food3/src/pages/Carrinho/Carrinho.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import { Titulo } from './style'
-import { Endereco } from './style'
+import { DescriptionEndereco, Titulo } from './style'
+import { Endereco, TitleEndereco } from './style'
 import CarrinhoVazio from '../../components/CarrinhoVazio/CarrinhoVazio'
 import CarrinhoComItens from '../../components/CarrinhoComItens/CarrinhoComItens'
 import Header from '../../components/Header/Header'
@@ -13,8 +13,8 @@ export default function Carrinho() {
             <Header title = 'Meu Carrinho'/>
 
             <Endereco>
-                <h3>Endereço de entrega</h3>
-                <h3>Rua Alessandra Vieira, 42</h3>
+                <TitleEndereco>Endereço de entrega</TitleEndereco>
+                <DescriptionEndereco><b>Rua Alessandra Vieira, 42</b></DescriptionEndereco>
             </Endereco>
             {/* <CarrinhoVazio/> */}
             <CarrinhoComItens/>
