+++ labe-food3/src/pages/PerfilPage/EditEndereco/EditEndereco.js
@@ -2,15 +2,15 @@ import React from "react";
 import { useNavigate } from "react-router-dom";
 import { useForm } from "../../../hooks/useForm";
 import TextField from "@mui/material/TextField";
-import { ButtonGo, Title } from "./styled";
+import { ButtonGo, Title, ChevronLeftIconE } from "./styled";
 import Stack from "@mui/material/Stack";
 import { goPerfil } from "../../../routes/coordinator";
 import { BASE_URL } from "../../../constants/urls";
 import axios from "axios";
-import Header from '../../../components/Header/Header'
 import { ScreenContainer, InputsContainer } from "./styled";
-import ChevronRightIcon from '@mui/icons-material/ChevronRight'
+import ChevronLeftIcon from '@mui/icons-material/ChevronLeft'
 import { goBack } from "../../../routes/coordinator";
+import Header from '../../../components/Header/Header'
 
 const Endereco = () => {
 
@@ -51,7 +51,7 @@ const Endereco = () => {
 
   return (
     <div>
-    <Header title ='Editar Perfil' back = 'true'  onClick={()=>goBack(navigate)}  icon={<ChevronRightIcon sx={{ fontSize: 36 }} />}/>
+    <Header title ={<ChevronLeftIconE onClick={()=>goBack(navigate)} sx={{ fontSize: 36 }} />} /> 
    
     <ScreenContainer> 
       <form onSubmit={onSubmitAdress}>
